requirements.txt
python-telegram-bot==21.4
pyrogram==2.0.106
pytgcalls==3.0.0.dev24
yt-dlp==2024.4.9
tgcrypto==1.2.5
